Move conceptMonitoring in separate file.
