+++ api/LMPlatform.UI/Services/Concept/IConceptService.cs
@@ -1,6 +1,4 @@
-﻿using LMPlatform.Models;
-using LMPlatform.UI.Services.Modules.Concept;
-using System.Collections.Generic;
+﻿using LMPlatform.UI.Services.Modules.Concept;
 using System.ServiceModel;
 using System.ServiceModel.Web;
 using static LMPlatform.UI.Services.Concept.ConceptService;
